my silly bugs removed !;)
