+++ controllers/users/getCurrentUser.js
@@ -6,6 +6,6 @@ export async function getCurrentUser(req, res) {
       return res.status(404).json({ message: "User not found" });
     }
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    return res.status(500).json({ message: error.message });
   }
 }
