+++ server.js
@@ -4,9 +4,9 @@ import mongoose from "mongoose";
 import dotenv from "dotenv";
 dotenv.config();
 // Connection string do twojej bazy danych MongoDB w MongoDB Atlas
-const dbConnectionURI = process.env.DB_URL;
-//const dbConnectionURI =
-("mongodb+srv://mielnikmagdalena:3kMJIYHF3nYNLhGT@cluster0.5lgwxxf.mongodb.net/db-contacts");
+//const dbConnectionURI = process.env.DB_URL;
+const dbConnectionURI =
+  "mongodb+srv://mielnikmagdalena:3kMJIYHF3nYNLhGT@cluster0.5lgwxxf.mongodb.net/db-contacts";
 // Ustanawianie połączenia
 mongoose.connect(dbConnectionURI, {
   useNewUrlParser: true,
